% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psica.R
\name{plot.psicaTree}
\alias{plot.psicaTree}
\title{Plots a PSICA tree.}
\usage{
\method{plot}{psicaTree}(x, type = 1, ...)
}
\arguments{
\item{x}{the PSICA tree to be plotted.}

\item{type}{If type=1 a default plot showing the probability of each treatment to be the best are
displayed, if type=2 boxplots of the effect values are displayed in each terminal node.}

\item{...}{Other parameter passed to the generic plot function}
}
\description{
A plot is produced that shows a PSICA tree in which the terminal nodes
contain the probabilities that one treatment is better than the other
ones as well as a label containing the possible best treatments
}
