% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psica.R
\name{predict.psicaTree}
\alias{predict.psicaTree}
\title{Makes predictions for PSICA tree.}
\usage{
\method{predict}{psicaTree}(object, prob = FALSE, ...)
}
\arguments{
\item{object}{the PSICA tree to be predicted}

\item{prob}{should the matrix or class probabilities be predicted (TRUE)
or the classes themselves (FALSE)}

\item{...}{further parameters to be passed to 'rpart' object.}
}
\description{
Makes predictions for PSICA tree.
}
