% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psica.R
\name{print.psicaTree}
\alias{print.psicaTree}
\title{Prints a PSICA tree.}
\usage{
\method{print}{psicaTree}(x, ...)
}
\arguments{
\item{x}{the PSICA tree to be printed}

\item{...}{Other parameter passed to the generic print function}
}
\description{
Prints a PSICA tree.
}
