% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psica.R
\name{prune.psicaTree}
\alias{prune.psicaTree}
\title{Prunes a PSICA tree.}
\usage{
\method{prune}{psicaTree}(pT, cp = 0.001, ...)
}
\arguments{
\item{pT}{the PSICA tree to be pruned.}

\item{cp}{Cost complexity parameter that defines how much the tree
must be pruned. See rpart::control() for description of the
cost-complexity parameter.}

\item{...}{further parameters to be passed to 'rpart' object.}
}
\description{
Prunes a PSICA tree.
}
