% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psica.R
\name{snip.psicaTree}
\alias{snip.psicaTree}
\title{Prunes desired nodes in a PSICA tree.}
\usage{
\method{snip}{psicaTree}(pT, indices)
}
\arguments{
\item{pT}{the PSICA tree to be pruned.}

\item{indices}{vector of indices of the nodes. All contents below the nodes with these
indices will be pruned.}
}
\description{
Prunes desired nodes in a PSICA tree.
}
